created a dog card template component
