+++ src/Components/CreateDog/CreateDogList.styles.scss
@@ -1,14 +1,5 @@
 @import "../../utils/_variables.scss";
 
-.card {
-    width: 300px;
-    margin: 10px;
-    box-shadow: rgba(99, 99, 99, 0.2) 0px 2px 8px 0px;
-}
-.card-img-top {
-    width: 100% !important;
-    height: 250px;
-}
 
 .cardGrid {
     display: flex;
