+++ src/Components/DogCardTemplate/DogCardTemplate.styles.scss
@@ -1,0 +1,15 @@
+.card {
+    width: 300px;
+    margin: 10px;
+    box-shadow: rgba(99, 99, 99, 0.2) 0px 2px 8px 0px;
+}
+.card-img-top {
+    width: 100% !important;
+    height: 250px;
+}
+
+@media screen and (max-width: 1200px) {
+    .card {
+        width: 200px;
+    }
+}
